# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Buddy360 API",
    "description": "Backend API for the Buddy360 parenting-companion app. Provides authentication, child-profile management, LLM-powered guidance, and audio transcription.",
    "contact": {
      "name": "Buddy360 Engineering",
      "email": "prince.kumar@pearson.com"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
  "paths": {
...
    "/api/v1/children": {
      "get": {
        "tags": [
          "children"
        ],
        "summary": "List Children",
        "description": "List all children linked to the authenticated user's account.",
        "operationId": "list_children_api_v1_children_get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "created_date",
                    "-created_date",
                    "name",
                    "-name"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "-created_date",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChildResponse"
                  },
                  "title": "Response List Children Api V1 Children Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "children"
        ],
        "summary": "Create Child",
        "description": "Add a new child profile to the authenticated user's account (maximum 10).",
        "operationId": "create_child_api_v1_children_post",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/children/{child_id}": {
      "get": {
        "tags": [
          "children"
        ],
        "summary": "Get Child",
        "description": "Retrieve a single child profile by ID.",
        "operationId": "get_child_api_v1_children__child_id__get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Child Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "children"
        ],
        "summary": "Update Child",
        "description": "Update details of an existing child profile.",
        "operationId": "update_child_api_v1_children__child_id__patch",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Child Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "children"
        ],
        "summary": "Delete Child",
        "description": "Soft-delete a child profile. The profile is hidden immediately but retained for 30 days so accidental deletions can be recovered. Associated data (growth areas, observations, etc.) is preserved during the retention window and purged by a scheduled hard-delete job after expiry.",
        "operationId": "delete_child_api_v1_children__child_id__delete",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Child Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/children/{child_id}/progress/{flag}": {
      "post": {
        "tags": [
          "children"
        ],
        "summary": "Mark Journey Progress",
        "description": "Mark a Personality Journey progression flag as done. One-way (false\u2192true), one flag per call, gated on the previous step in the chain already being true \u2014 this is the only way to set these flags; PATCH /children/{child_id} silently ignores them.",
        "operationId": "mark_journey_progress_api_v1_children__child_id__progress__flag__post",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Child Id"
            }
          },
          {
            "name": "flag",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "onboarding_profile_completed",
                "discover_completed",
                "transform_visited",
                "release_visited",
                "connect_visited"
              ],
              "type": "string",
              "title": "Flag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/children/{child_id}/avatar/presign": {
      "post": {
        "tags": [
          "children"
        ],
        "summary": "Presign Child Avatar",
        "description": "Generate a presigned S3 PUT URL for uploading a child avatar photo. The client uploads directly to S3 using this URL, then PATCHes the child with the returned avatar_url.",
        "operationId": "presign_child_avatar_api_v1_children__child_id__avatar_presign_post",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Child Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarPresignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarPresignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvatarPresignRequest": {
        "properties": {
          "content_type": {
            "type": "string",
            "maxLength": 100,
            "title": "Content Type"
          }
        },
        "type": "object",
        "required": [
          "content_type"
        ],
        "title": "AvatarPresignRequest"
      },
      "AvatarPresignResponse": {
        "properties": {
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "avatar_url": {
            "type": "string",
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "upload_url",
          "avatar_url"
        ],
        "title": "AvatarPresignResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}